% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{batch.adsa}
\alias{batch.adsa}
\title{Calculates reallocated batch size or new input location for ADSA}
\usage{
batch.adsa(fit, r_seq, xtest, xt_dens, x_new, r0, nugget, method)
}
\arguments{
\item{fit:}{gp/tp fit}

\item{r_seq:}{batch size vector for existing inputs}

\item{xtest:}{testing points to compare reallocation and adding a new inputs}

\item{xt_dens:}{density of xtest}

\item{x_new:}{new input location selected by one EI criteria}

\item{r0:}{total number of new simulations}

\item{nugget:}{the noise variance to compute the ALC factor}

\item{method:}{"km" or "trainkm" or "hetgp" or "homtp"}
}
\description{
ADSA for Adaptive Batching
}
