% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{batch.ddsa}
\alias{batch.ddsa}
\title{Calculates reallocated batch size for DDSA}
\usage{
batch.ddsa(fit, r_seq, xtest, xt_dens, r0, method)
}
\arguments{
\item{fit}{gp/tp fit}

\item{r_seq}{batch size vector for existing inputs}

\item{xtest}{testing points to compare reallocation and adding a new inputs}

\item{xt_dens}{density of xtest}

\item{r0}{total number of new simulations to add}

\item{method}{"km" or "trainkm" or "hetgp" or "homtp"}
}
\description{
DDSA for Adaptive Batching
}
\references{
M. Ludkovski, X. Lyu (2020+) Adaptive Batching for Gaussian Process Surrogates with Application 
in Noisy Level Set Estimation, <http://arxiv.org/abs/2003.08579>
}
\seealso{
\code{\link[mlOSP]{osp.seq.batch.design}}
}
\author{
Xiong Lyu
}
