% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{batch.mlb}
\alias{batch.mlb}
\title{Calculates weights for batch size in MLB}
\usage{
batch.mlb(objSd_at_x_optim, r_cand, nugget, gamma)
}
\arguments{
\item{objSd_at_x_optim:}{posterior standard deviation of the response at the selected new input}

\item{r_cand:}{candidate batch size vector}

\item{nugget:}{the noise variance to compute the ALC factor}

\item{gamma:}{threshold compared with sd}
}
\description{
MLB for Adaptive Batching
}
