% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{CalcOverhead}
\alias{CalcOverhead}
\title{Calculates c_over in ABSUR}
\usage{
CalcOverhead(theta0, theta1, theta2, n)
}
\arguments{
\item{theta0, theta1, theta2}{parameters in linear regression}

\item{n}{current design size}
}
\description{
ABSUR overhead
}
\details{
compute the estimated overhead in ABSUR
}
