% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{call.payoff}
\alias{call.payoff}
\title{Arithmetic average Call payoff \eqn{(mean(x)-K)_+}}
\usage{
call.payoff(x, model)
}
\arguments{
\item{x}{is a matrix of asset prices of dimension N * \code{model$dim}}

\item{model}{list containing model params. Uses \code{model$K} as the Put strike.}
}
\description{
Basket Call option on average asset price
}
\details{
arithmetic basket for d-dim x)
}
