% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{cf.absur}
\alias{cf.absur}
\title{Calculates weights for location and batch size in ABSUR}
\usage{
cf.absur(objMean, objSd, nugget, r_cand, overhead, t0)
}
\arguments{
\item{objMean:}{predicted mean response}

\item{objSd:}{posterior standard deviation of the response}

\item{nugget:}{the noise variance to compute the ALC factor}

\item{r_cand:}{candidate batch size vector}

\item{overhead:}{estimated computation overhead in GP}

\item{t0:}{overhead for individual simulation}
}
\description{
ABSUR for Adaptive Batching
}
