% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{cf.csur}
\alias{cf.csur}
\title{Compute reduction in contour-distance}
\usage{
cf.csur(objMean, objSd, nugget)
}
\arguments{
\item{objMean}{predicted mean response}

\item{objSd}{posterior standard deviation of the response}

\item{nugget}{the noise variance to compute the ALC factor}
}
\description{
cSUR for Contour Finding based on Ludkovski (2018)
}
\details{
compute the change in ZC = sd*(1-sqrt{(nugget)})/sqrt{(nugget + sd^2)}
}
\references{
Mike Ludkovski, Kriging Metamodels and Experimental Design for Bermudan Option Pricing
 Journal of Computational Finance, 22(1), 37-77, 2018
}
\seealso{
[osp.seq.design]
}
\author{
Mike Ludkovski
}
