% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{digital.put.payoff}
\alias{digital.put.payoff}
\title{geometric digital Put payoff}
\usage{
digital.put.payoff(x, K = 10)
}
\arguments{
\item{x}{is a vector of asset prices}

\item{K}{is the strike}
}
\description{
geometric digital Put payoff
}
