% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{forest.impulse}
\alias{forest.impulse}
\title{Compute intervention function for the Faustmann forest rotation problem}
\usage{
forest.impulse(cur_x, model, fit, ext = FALSE)
}
\arguments{
\item{cur_x}{Set of inputs where to compute the intervention function.
Should be a n x 1 vector}

\item{model}{a list containing all model parameters,
including \code{model$impulse.fixed.cost} for the constant cost of any impulse}

\item{fit}{Object containing the one-step-ahead functional approximator for V(k,x)}

\item{ext}{logical flag (default is FALSE) whether to return extended information}
}
\description{
Compute intervention function for the Faustmann forest rotation problem
}
\details{
Calculates the intervention operator for a 1-D impulse control problem
arising in the Faustmann forest rotation setup. In that case, the impulse target level
is fixed at zero (or \code{model$impulse.target}) and the impulse value is x-fixed.cost-target
Calls \code{ospPredict} on \code{fit} to find that
}
