% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{forward.sim.policy}
\alias{forward.sim.policy}
\title{Forward simulation based on a sequence of emulators}
\usage{
forward.sim.policy(x, M, fit, model, offset = 1, compact = TRUE,
  use.qv = FALSE)
}
\arguments{
\item{x}{is a matrix of starting values

if input \code{x} is a list, then use the grids specified by x}

\item{M}{is number of time steps to forward simulate}

\item{fit}{is a list of fitted emulators that determine the stopping classifiers to be used}

\item{use.qv}{boolean to indicate whether to plug-in continuation value for all remaining paths at the last time-step
default is set to \code{FALSE}}
}
\value{
a list containing:
\itemize{
 \item \code{payoff} is the resulting payoff NPV from t=0
 \item \code{fvalue[i]} is a list of resulting payoffs (on paths still not stopped) NPV from t=i
 \item \code{tau} are the times when stopped
 \item \code{sims} is a list; \code{sims[[i]]} are the forward x-values of paths at t=i (those not stopped yet)
\code{nsims} number of total 1-step simulations performed
}
}
\description{
Simulate h(X_tau) using FIT (can be a dynaTree, smooth.spline, MARS or RandomForest or RVM or hetGP)
}
\details{
Should be used in conjuction with the \code{osp.} functions that build the emulators. Also called
internally from \code{\link{osp.fixed.design}}
}
