% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\docType{data}
\name{int300_3d}
\alias{int300_3d}
\title{Initial design for the 3D Bermudan max-Call example}
\format{
An array with 300 rows and 3 columns:
\describe{
  \item{row}{unique designs}
  \item{columns}{values of corresponding x_1, x_2 and x_3 coordinates}
}
}
\usage{
int300_3d
}
\description{
A dataset containing 300 initial locations to be used with \code{osp.seq.design}. 
The max Call has strike K=100 and i.i.d assets, so the experimental design is roughly symmetric.
The design is space-filling a hyper-rectangular in-the-money region
}
\keyword{datasets}
