% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\docType{data}
\name{int_2d}
\alias{int_2d}
\title{Initial design for the 2D Bermudan Put example}
\format{
An array with 20 rows and 2 columns:
\describe{
  \item{row}{unique designs}
  \item{columns}{values of corresponding x_1 and x_2 coordinates}
}
}
\usage{
int_2d
}
\description{
A dataset containing 20 initial locations to be used with \code{osp.seq.design}. 
The Put has strike K=40 and i.i.d assets, so the experimental design is roughly symmetric.
The design is space-filling a triangular in-the-money region
}
\keyword{datasets}
