% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{lin.impulse}
\alias{lin.impulse}
\title{Compute intervention function for an impulse problem wth linear impulse costs}
\usage{
lin.impulse(cur_x, model, fit, ext = FALSE)
}
\arguments{
\item{cur_x}{Set of inputs where to compute the intervention function
Should be a n x 1 vector}

\item{model}{a list containing all model parameters. 
In particular must have \code{model$impulse.fixed.cost} for the constant cost of any impulse}

\item{fit}{Object containing the one-step-ahead functional approximator for V(k,x)}

\item{ext}{logical flag (default is FALSE) whether to return extended information}
}
\description{
Compute intervention function for an impulse problem wth linear impulse costs
}
\details{
Calculates the intervention operator for a 1-D impulse control problem.
Assumes linear impulse costs with slope=1. This means that the optimal impulse 
target level is independent of current state x and is characterized by the location
where the gradient of fitted value function is equal to 1.
Calls \code{ospPredict} on \code{fit} to find that
}
