% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{maxCall}
\alias{maxCall}
\title{Max Call payoff}
\usage{
maxCall(x, K = 100)
}
\arguments{
\item{x}{is a vector of asset prices}

\item{K}{is the strike}

\item{x:}{asset prices}
}
\description{
Max Call payoff
}
