% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{mini.put.payoff}
\alias{mini.put.payoff}
\title{Min Put payoff}
\usage{
mini.put.payoff(x, model)
}
\arguments{
\item{x}{Matrix of asset prices  with \code{model$dim} columns and N rows}

\item{model}{list containing model params. Uses \code{model$K} as the Put strike.}
}
\description{
Multivariate min Put
}
