% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{minPut}
\alias{minPut}
\title{Min Put payoff}
\usage{
minPut(x, K = 40)
}
\arguments{
\item{x}{is a vector of asset prices}

\item{K}{is the strike}

\item{x:}{asset prices}
}
\description{
Min Put payoff
}
