% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospPiecewiseBW.R
\name{osp.probDesign.piecewisebw}
\alias{osp.probDesign.piecewisebw}
\title{Longstaff Schwartz Algorithm using the Bouchard-Warin method

Uses the Bouchard-Warin recursive partitioning to create N-d trees
for local linear regression fits. Each tree node contains N/model$nChildren^model$dim inputs.}
\usage{
osp.probDesign.piecewisebw(N, model, test.paths = NULL, verb = 0)
}
\arguments{
\item{N}{the number of forward training paths}

\item{model}{a list defining all model parameters. Must contain the following fields:
\cr \code{T, dt, dim, nChildren},
       \code{sim.func, x0, r, payoff.func}}

\item{test.paths}{(optional) a list containing out-of-sample paths to obtain a price estimate}

\item{verb}{if specified, produces plots of the 1-dim fit every \code{verb} time-steps
[default is zero, no plotting]}
}
\value{
a list with the following fields:
\itemize{
\item \code{price} is the scalar optimal reward;
\item \code{tau} is a vector of stopping times over in-sample paths;
\item \code{test} is a vector of out-of-sample pathwise rewards;
\item \code{val} is a vector of in-sample pathwise rewards
\item \code{timeElapsed} total running time based on \code{Sys.time}
}
}
\description{
Longstaff Schwartz Algorithm using the Bouchard-Warin method

Uses the Bouchard-Warin recursive partitioning to create N-d trees
for local linear regression fits. Each tree node contains N/model$nChildren^model$dim inputs.
}
\details{
Calls \link{treeDivide.BW} to create the equi-probable partitions.
Must have N/model$nChildren^model$dim as an integer.
}
\examples{
set.seed(1)
modelSV5 <- list(K=100,x0=c(90, log(0.35)),r=0.0225,div=0,sigma=1,
   T=50/252,dt=1/252,svAlpha=0.015,svEpsY=1,svVol=3,svRho=-0.03,svMean=2.95,
   eulerDt=1/2520, dim=2,sim.func=sim.expOU.sv,nChildren=10,payoff.func=sv.put.payoff)
putPr <- osp.probDesign.piecewisebw(20000,modelSV5)
putPr$price
  # get [1] 17.30111
}
\references{
Bruno Bouchard and Xavier Warin. Monte-Carlo valorisation of American options: facts and new
algorithms to improve existing methods. In R. Carmona, P. Del Moral, P. Hu, and N. Oudjane, editors,
Numerical Methods in Finance, volume 12 of Springer Proceedings in Mathematics. Springer, 2011.
}
