% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospSeqBatchDesignSimplified.R
\name{osp.seq.batch.design.simplified}
\alias{osp.seq.batch.design.simplified}
\title{Adaptive Batch design for optimal stopping (simplified version)}
\usage{
osp.seq.batch.design.simplified(model, method = "km")
}
\arguments{
\item{model}{a list containing all model parameters}

\item{method}{\code{km} to select the GP emulator to apply}
}
\value{
a list containing:
\itemize{
\item \code{fit} a list of fitted response surfaces
\item \code{ndesigns}: number of design size k_T
\item \code{batches}: matrix of replications r_i
}
}
\description{
Adaptive Batch design for optimal stopping (simplified version)
}
\details{
Implements the adaptive batching strategy defined in batch.heuristic with model defined in method.
}
