% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{ospPredict}
\alias{ospPredict}
\title{Forward simulation of a swing payoff based on a sequence of emulators}
\usage{
ospPredict(myFit, myx, model)
}
\arguments{
\item{myFit}{fitted emulators of any type supported by osp.prob.design}

\item{myx}{inputs to predict at}

\item{model}{List containing all model parameters.}
}
\value{
prediction of myFit evaluated at myx
}
\description{
Evaluate the \code{fit} emulator at input \code{myx}
}
\details{
Generally internal to other mlOSP functions
}
