% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{plot_style}
\alias{plot_style}
\title{GGplot style}
\usage{
plot_style(base_size = 14, base_family = "Helvetica", ...)
}
\arguments{
\item{base_size}{is the font size}

\item{base_family}{is the font style}

\item{...}{is for the parameters to specialize the ggplot style}
}
\description{
GGplot style
}
