% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{plt.1d.fit}
\alias{plt.1d.fit}
\title{Make a 3-panel plot of the 1-d fit from \code{mc.put.adaptree}}
\usage{
plt.1d.fit(
  obj,
  x.gr = seq(28, 40, len = 500),
  new.fig = F,
  show.ci = "none",
  heuristic = "zc",
  al.weights = NULL,
  obj2 = NULL,
  ...
)
}
\arguments{
\item{obj}{a list containing the fitted emulator}

\item{x.gr}{locations to use for the predictions}

\item{new.fig}{flag whether to overplot or not}

\item{show.ci}{can be 'none', 'km', 'dt-vmean', 'dt-df'}

\item{heuristic}{calls the corresponding qEI function}

\item{al.weights}{EI weights}

\item{obj2}{(experimental) a reference \code{km} emulator to compare against}
}
\description{
Make a 3-panel plot of the 1-d fit from \code{mc.put.adaptree}
}
