% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{plt.2d.surf}
\alias{plt.2d.surf}
\title{Visualize a 2D emulator + stopping region}
\usage{
plt.2d.surf(
  fit,
  x = seq(31, 43, len = 201),
  y = seq(31, 43, len = 201),
  ub = 1e+06,
  show.var = FALSE,
  only.contour = FALSE,
  contour.col = "red",
  bases = NULL
)
}
\arguments{
\item{fit}{a fitted emulator. can be any of the types supported by \code{\link{forward.sim.policy}}}

\item{x, y}{locations to use for the \code{predict()} functions. Default is a 200x200 fine grid.
Passed to \code{expand.grid}}

\item{ub}{clip the surface with an upper bound  to see the zero-contour better}

\item{show.var}{if \code{TRUE} then plot posterior surrogate variance instead of surrogate mean [default = FALSE]
This only works for \code{km} and \code{het/homGP/homTP} objects}

\item{only.contour}{-- just add the zero-contour, no raster plot (uses \code{contour}, no ggplot)}

\item{contour.col}{(default is "red") -- color of the zero contour}

\item{bases}{(only used for lm objects)}
}
\value{
a ggplot handle for the created plot.
}
\description{
Two-dimensional raster+contour+point plot of an mlOSP emulator at a single time step.
}
\details{
Uses the raster plot from \pkg{ggplot2}. For GP-based objects, also shows the unique design
sites via geom_point. See \code{\link{plt.2d.surf.batch}} for a similar plot 
for \code{osp.seq.batch.design} emulators.
}
