% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{put.payoff}
\alias{put.payoff}
\title{Arithmetic basket Put for d-dim x
 @title American Put payoff}
\usage{
put.payoff(x, model)
}
\arguments{
\item{x}{is a matrix of asset prices of dimension N * \code{model$dim}}

\item{model}{list containing model params. Uses \code{model$K} as the Put strike.}
}
\description{
Arithmetic basket Put for d-dim x
 @title American Put payoff
}
\details{
in more than 1D, the prices are averaged and maxed with zero.
}
