% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{r.reallocate}
\alias{r.reallocate}
\title{Calculates reallocated batch size}
\usage{
r.reallocate(L, K, xt_dens, r_seq, r0)
}
\arguments{
\item{L:}{lower triangle of cholesky decomposition of covariance matrix}

\item{K:}{covariance matrix}

\item{xt_dens:}{density of xtest}

\item{r_seq:}{batch size vector for existing inputs}

\item{r0:}{total number of new simulations}
}
\description{
New batch size calculator
}
