% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.bm}
\alias{sim.bm}
\title{Simulate 1D arithmetic Brownian Motion}
\usage{
sim.bm(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values (matrix of size N x model$dim)}

\item{model}{a list containing all the other parameters, including volatility \code{model$sigma},
interest rate \code{model$r} and continuous dividend yield \code{model$div}.}

\item{dt}{is the step size in time. Defaults to \code{model$dt}}

\item{model$drift}{for the drift term}
}
\description{
Simulate 1D arithmetic Brownian Motion
}
