% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.expOU.sv}
\alias{sim.expOU.sv}
\title{Simulate an exp-OU stoch volatility model with 1 or 2 vol factors}
\usage{
sim.expOU.sv(x0, model, dt = model$dt, useEuler = F)
}
\arguments{
\item{x0}{should have 2 or 3 columns}

\item{model}{contains all the other parameters}

\item{dt}{is the step size}
}
\description{
Simulate an exp-OU stoch volatility model with 1 or 2 vol factors
}
\section{Usage}{
  need the following fields in model: svMean (mean-rev level), svAlpha (men-rev strength),
svEpsY (fast scaling param), svVol (vol-vol), svRho (corr)
for 2-factor also need: svMeanZ (slow scale mean-rev level), svAlphaZ (mean-rev strength), svDeltaZ (slow
scaling param)
svVolZ (Z vol), svRhoZ (corr between Z and S), svRhoYZ( corr between the 2 factors)
}

