% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.gbm.asian}
\alias{sim.gbm.asian}
\title{Simulate 1D Brownian Motion for Asian Options
first column is t, second column is S_t
third column is A_t (arithmetic average)
fourth column is tilde{A}_t (geometric average)}
\usage{
sim.gbm.asian(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values (vector)}

\item{model}{contains all the other parameters, including volatility \code{model$sigma}
interest rate \code{model$r} and continuous dividend yield \code{model$div}}

\item{dt}{is the step size}
}
\description{
Simulate 1D Brownian Motion for Asian Options
first column is t, second column is S_t
third column is A_t (arithmetic average)
fourth column is tilde{A}_t (geometric average)
}
