% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.gbm.cor}
\alias{sim.gbm.cor}
\title{Simulate paths of correlated GBM with a constant correlation}
\usage{
sim.gbm.cor(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values (vector)}

\item{model}{contains all the other parameters.
In particular, need \code{model$r, model$rho, model$sigma, model$div, model$dim}
Note that \code{model$sigma} is the \strong{volatility} parameter (scalar)}

\item{dt}{is the step size}
}
\description{
Simulate paths of correlated GBM with a constant correlation
}
\details{
Simulate correlated multivariate Geometric Brownian motion
with a given \code{model$rho} and \strong{identical} \code{model$sigma}'s
}
