% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.gbm}
\alias{sim.gbm}
\title{Simulate paths of Geometric Brownian Motion with constant parameters}
\usage{
sim.gbm(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values (matrix of size N x model$dim)}

\item{model}{a list containing all the other parameters, including volatility \code{model$sigma},
interest rate \code{model$r} and continuous dividend yield \code{model$div}.}

\item{dt}{is the step size in time. Defaults to \code{model$dt}}
}
\value{
a vector of same dimensions as x0
}
\description{
Simulate paths of Geometric Brownian Motion with constant parameters
}
\details{
Simulate from \eqn{p(X_t|X_{t-1})}.
Use log-normal transition density specified by the \code{model} parameters
}
