% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.logOU_Discrete}
\alias{sim.logOU_Discrete}
\title{Simulate from 1-D discretized exponential OU
See Bender (SIFIN 2011). Only works in 1D}
\usage{
sim.logOU_Discrete(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values}

\item{model}{contains all the other parameters}

\item{dt}{is the step size}
}
\description{
Simulate from 1-D discretized exponential OU
See Bender (SIFIN 2011). Only works in 1D
}
\details{
requires
\itemize{
\item  model$rho -- similar to mean-reversion rate, should be close to 1
\item model$mu -- mean-reversion level
\item model$sigma -- volatility
}
}
