% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.ouExp}
\alias{sim.ouExp}
\title{Simulate from exponential OU}
\usage{
sim.ouExp(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values}

\item{model}{contains all the other parameters}

\item{dt}{is the step size}
}
\description{
Simulate from exponential OU
}
