% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.price.and.capacity}
\alias{sim.price.and.capacity}
\title{Simulate 2D process for price and capacity}
\usage{
sim.price.and.capacity(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values (matrix of size N x model$dim)}

\item{model}{a list containing all the other parameters, including volatility \code{model$sigma},
interest rate \code{model$r} and continuous dividend yield \code{model$div}.}

\item{dt}{is the step size in time. Defaults to \code{model$dt}}
}
\description{
Simulate 2D process for price and capacity
}
\details{
first column is P_t price, second column is C_t capacity
}
