% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{sv.put}
\alias{sv.put}
\title{Basket Put with stochastic volatility}
\usage{
sv.put(x, K = 40)
}
\arguments{
\item{x}{is a vector of asset prices}

\item{K}{is the strike}

\item{x:}{First coordinate is the asset price, the rest are vol factors}
}
\description{
Put payoff for a stoch vol model
}
\details{
Uses K-x[,1], ignores other coordinates
}
