% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{batch.adsa}
\alias{batch.adsa}
\title{Calculates reallocated batch size or new input location for Adaptive Design with Sequential Allocation}
\usage{
batch.adsa(fit, r_seq, xtest, xt_dens, x_new, r0, nugget, method)
}
\arguments{
\item{fit:}{gp/tp fit}

\item{r_seq:}{batch size vector for existing inputs}

\item{xtest:}{testing points to compare reallocation and adding a new inputs}

\item{xt_dens:}{density of xtest}

\item{x_new:}{new input location selected by one EI criteria}

\item{r0:}{total number of new simulations}

\item{nugget:}{the noise variance to compute the ALC factor}

\item{method:}{\code{km} or \code{trainkm} or \code{hetgp} or \code{homtp}}
}
\value{
a list containing
\itemize{
\item \code{xoptim}: new design input (NULL if re-allocation is chosen)
\item \code{roptim}: added replications (a scalar r0 if new input chosen, 
a vector containing the re-allocation amounts otherwise
}
}
\description{
ADSA for Adaptive Batching
}
\references{
M. Ludkovski, X. Lyu (2020+) Adaptive Batching for Gaussian Process Surrogates with Application 
in Noisy Level Set Estimation, http://arxiv.org/abs/2003.08579
}
\seealso{
\link{\code{osp.seq.batch.design}}
}
\author{
Xiong Lyu
}
