% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{batch.mlb}
\alias{batch.mlb}
\title{Multi-Level Batching Heuristic}
\usage{
batch.mlb(objSd_at_x_optim, r_cand, nugget, gamma)
}
\arguments{
\item{objSd_at_x_optim}{posterior standard deviation of the response at the selected new input}

\item{r_cand}{candidate batch size vector}

\item{nugget}{the noise variance to compute the ALC factor}

\item{gamma}{threshold compared with sd}
}
\value{
list containing:
 \itemize{ 
 \item \code{roptim}: new replication count
 \item \code{gamma}: new gamma variable
 }
}
\description{
Calculates weights for batch size in MLB when called from \link{osp.seq.batch.design}
}
\references{
M. Ludkovski, X. Lyu (2020+) Adaptive Batching for Gaussian Process Surrogates with Application 
in Noisy Level Set Estimation, http://arxiv.org/abs/2003.08579
}
\seealso{
\code{\link[mlOSP]{osp.seq.batch.design}}
}
