% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{call.payoff}
\alias{call.payoff}
\title{Arithmetic average Call payoff}
\usage{
call.payoff(x, model)
}
\arguments{
\item{x}{is a vector of asset prices}
}
\description{
Call
}
\details{
arithmetic basket for d-dim x)
}
