% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{call.payoff}
\alias{call.payoff}
\title{American Call payoff}
\usage{
call.payoff(x, K = 40)
}
\arguments{
\item{x}{is a vector of asset prices}

\item{K}{is the strike}
}
\description{
American Call payoff
}
\details{
arithmetic basket for d-dim x)
}
