% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{cf.el}
\alias{cf.el}
\title{Expected Loss for Contour Finding}
\usage{
cf.el(objMean, objSd)
}
\arguments{
\item{obj}{must be a DynaTree or a list with an sd field}
}
\description{
Compute expected loss using the optimal stopping loss function.
}
