% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{cf.sur}
\alias{cf.sur}
\title{Compute EI for Contour Finding using the ZC-SUR formula}
\usage{
cf.sur(objMean, objSd, nugget)
}
\arguments{
\item{objMean:}{predicted mean response}

\item{objSd:}{posterior standard deviation of the response}
}
\description{
SUR (Stepwise Uncertainty Reduction) acquisition function for Contour Finding
}
\details{
compute the change in ZC = sd*(1-sqrt{(nugget)})/sqrt{(nugget + sd^2)}
}
