% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{cf.tMSE}
\alias{cf.tMSE}
\title{Maximum Contour Uncertainty criterion}
\usage{
cf.tMSE(objMean, objSd, seps = 0.05)
}
\arguments{
\item{objMean:}{predicted mean response}

\item{objSd:}{posterior standard deviation of the response}

\item{seps:}{epsilon in the tMSE formula}
}
\description{
tMSE for Contour Finding
}
\details{
compute predictive density at the contour, smoothed by seps
}
