% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{geom.put.payoff}
\alias{geom.put.payoff}
\title{geometric basket Put}
\usage{
geom.put.payoff(x, K = 40)
}
\arguments{
\item{x}{is a vector of asset prices}

\item{K}{is the strike}
}
\description{
geometric basket Put
}
