% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospProbDesign.R
\name{osp.fixed.design}
\alias{osp.fixed.design}
\title{Generic dynamic emulation of OSP with a non-sequential design}
\usage{
osp.fixed.design(
  model,
  input.domain = NULL,
  method = "km",
  inTheMoney.thresh = 0,
  stop.freq = model$dt
)
}
\arguments{
\item{model}{a list containing all the model parameters, see Details.}

\item{input.domain}{the domain of the emulator. Several options are available. Default in \code{NULL}
All the empirical domains rely on pilot paths generated using \code{pilot.nsims}>0 model parameter.
\itemize{
\item  NULL will use an empirical probabilistic design based on the pilot paths (default);
\item if a vector of length 2*model$dim then specifies the bounding rectangle
\item a single positive number, then build a bounding rectangle based on the 
\eqn{\alpha}-quantile of the pilot paths
\item a single negative number, then build a bounding rectangle based on the full range of the pilot paths
\item a vector specifies the precise design, used as-is (\emph{overrides design size})
}}

\item{method}{regression method to use (defaults to \code{km})
\itemize{
\item km: Gaussian process with fixed hyperparams  uses \pkg{DiceKriging} via \code{km} (default)
Must provide \code{km.cov} (vector of lengthscales) and \code{km.var} (process variance)
\item trainkm: GP w/trained hyperparams: uses \pkg{DiceKriging} via \code{km}
\item mlegp Local approximate GP from the \pkg{laGP}. Requires
\item homgp Homoskedastic GP: use \pkg{hetGP} with  \code{mleHomGP}
\item hetgp Heteroskedastic GP: use \pkg{hetGP} with \code{mleHetGP}
\item spline: Smoothing Splines, use \code{smooth.spline} (only in 1D). Requires number of 
knots via \code{model$nk} 
\item loess: Local Regression: use \code{loess} with \code{lo.span} parameter (only in 1D or 2D)
\item rvm: Relevance Vector Machine: uses \code{rvm} from \pkg{kernlab}. Can optionally provide
\code{rvm.kernel} parameter (default is 'rbfdot')
\item npreg: kernel regression using \pkg{np} package. Can optionally provide \code{np.kertype} 
 (default is "gaussian"); \code{np.regtype} (default is Linear-constant "lc"); \code{np.kerorder}
 (default kernel order is 2) and \code{np.bwtype} (default bandwidth type is "fixed")
\item lm: linear model from \pkg{stats}
}}

\item{inTheMoney.thresh}{which paths are kept, out-of-the-money is dropped.
Defines threshold in terms of \code{model$payoff.func}}

\item{stop.freq}{*experimental, currently disabled* frequency of stopping decisions (default is \code{model$dt}).
Can be used to stop less frequently.}
}
\value{
a list containing:
\itemize{
\item \code{fit} a list containing all the models generated at each time-step. \code{fit[[1]]} is the emulator
at \eqn{t=\Delta t}, the last one is \code{fit[[M-1]]} which is emulator for \eqn{T-\Delta t}.
\item \code{timeElapsed}: total running time based on \code{Sys.time}
}
}
\description{
RMC based on a batched non-adaptive design with a variety of regression methods
}
\details{
The design can be replicated through \code{batch.nrep} model parameter. Replication allows to use
nonparametric techniques which would be too expensive otherwise, in particular LOESS, GP and RVM.
All designs are restricted to in-the-money region, see \code{inTheMoney.thresh} parameter (modify at your own risk)
Thus, actual design size will be smaller than specified. By default, no forward evaluation is provided, i.e. the
method only builds the emulators. Thus, to obtain an actual estimate of the option price
combine with \code{\link{forward.sim.policy}}.
}
\examples{
set.seed(1)
model2d <- list(K=40,x0=rep(40,2),sigma=rep(0.2,2),r=0.06,div=0,
 T=1,dt=0.04,dim=2, sim.func=sim.gbm, payoff.func=put.payoff,pilot.nsims=1000,
 batch.nrep=100,kernel.family="matern5_2",N=400)
}
\author{
Mike Ludkovski
}
