% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{pegging.alg}
\alias{pegging.alg}
\title{Calculates reallocated batch size for ADSA}
\usage{
pegging.alg(r, U, r_seq)
}
\arguments{
\item{r}{total number of simulations}

\item{U}{weighted matrix for pegging algorithm}

\item{r_seq}{batch size vector for existing inputs}
}
\description{
Pegging algorithm for ADSA/DDSA
}
\references{
M. Ludkovski, X. Lyu (2020+) Adaptive Batching for Gaussian Process Surrogates with Application 
in Noisy Level Set Estimation, http://arxiv.org/abs/2003.08579
}
\author{
Xiong Lyu
}
