% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlOSP_utils.R
\name{plt.1d.fit}
\alias{plt.1d.fit}
\title{Make a 3-panel plot of the 1-d fit from mc.put.adaptree}
\usage{
plt.1d.fit(obj, x.gr = seq(28, 40, len = 500), new.fig = F,
  show.ci = "none", heuristic = "zc", al.weights = NULL,
  obj2 = NULL, ...)
}
\arguments{
\item{new.fig}{: whether to create a fresh panel}

\item{show.ci}{can be 'none', 'km', 'dt-vmean', 'dt-df'}

\item{heuristic:}{calls the corresponding qEI function}
}
\description{
Make a 3-panel plot of the 1-d fit from mc.put.adaptree
}
