% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{plt.2d.surf.with.batch}
\alias{plt.2d.surf.with.batch}
\title{Visualize 2D emulator + stopping region + batch amounts}
\usage{
plt.2d.surf.with.batch(
  fit,
  batch_size,
  x = seq(25, 50, len = 201),
  y = seq(25, 50, len = 201),
  show.var = FALSE,
  only.contour = FALSE,
  contour.col = "red"
)
}
\arguments{
\item{fit}{An emulator object. Can be any of the types supported by \code{\link[mlOSP]{forward.sim.policy}}}

\item{batch_size}{matrix of replication counts for each input generated by \code{\link[mlOSP]{osp.seq.batch.design}}}

\item{x, y}{locations to use for the \code{predict()} functions. Default is a 200x200 fine grid.
Passed to \code{expand.grid}}

\item{show.var}{if \code{TRUE} then plot posterior surrogate variance instead of 
surrogate mean (Default = FALSE)
This only works for \code{km} and \code{het/homGP} objects}

\item{only.contour}{-- just add the zero-contour, no geom_raster}

\item{contour.col}{(default is "red") -- color of the zero contour}
}
\description{
two-dimensional image+contour plot with replication counts for an \code{osp.seq.batch.design} fit
}
\author{
Xiong Lyu
}
