% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchDesign_utils.R
\name{plt.2d.surf.with.batch}
\alias{plt.2d.surf.with.batch}
\title{Visualize 2D emulator + stopping region + batch size}
\usage{
plt.2d.surf.with.batch(
  fit,
  batch_size,
  x = seq(25, 50, len = 201),
  y = seq(25, 50, len = 201),
  ub = 1e+06,
  show.var = FALSE,
  only.contour = FALSE,
  contour.col = "black",
  ...
)
}
\arguments{
\item{fit}{can be any of the types supported by \code{\link{forward.sim.policy}}}

\item{x, y}{locations to use for the \code{predict()} functions. Default is a 200x200 fine grid.
Passed to \code{expand.grid}}

\item{ub}{to create an upper bound to see the zero-contour better}

\item{show.var}{-- if \code{TRUE} then plot posterior surrogate variance instead of surrogate mean [default = FALSE]
This only works for \code{km} and \code{het/homGP} objects}

\item{only.contour}{-- just add the zero-contour, no quilt.plot}

\item{contour.col}{(default is "red") -- color of the zero contour}

\item{..}{-- pass additional options to \code{quilt.plot}}
}
\description{
two-dimensional image+contour plot with batch size
}
\details{
Uses the quilt.plot command from \pkg{fields}
}
