% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payoffs.R
\name{put.payoff}
\alias{put.payoff}
\title{Arithmetic basket Put for d-dim x
 @ title American Put payoff}
\usage{
put.payoff(x, K = 40)
}
\arguments{
\item{x}{is a vector of asset prices}

\item{K}{is the strike}
}
\description{
Arithmetic basket Put for d-dim x
 @ title American Put payoff
}
\details{
in more than 1D, the prices are averaged and maxed with zero.
}
