% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.gbm.cor}
\alias{sim.gbm.cor}
\title{Simulate paths of correlated GBM}
\usage{
sim.gbm.cor(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values}

\item{model}{contains all the other parameters}

\item{dt}{is the step size}
}
\description{
Simulate correlated multivariate Geometric Brownian motion
with a given \code{model$rho} and **identical** \code{model$sigma}'s
}
