% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.gbm.moving.ave}
\alias{sim.gbm.moving.ave}
\title{Simulate 1D Brownian Motion for Moving Average Asian Options
first column is S_t, other columns are lagged S_t's
the lags are in terms of dt
#' @inheritParams sim.gbm}
\usage{
sim.gbm.moving.ave(x0, model, dt = model$dt)
}
\description{
Simulate 1D Brownian Motion for Moving Average Asian Options
first column is S_t, other columns are lagged S_t's
the lags are in terms of dt
#' @inheritParams sim.gbm
}
