% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.gbm}
\alias{sim.gbm}
\title{Simulate paths of GBM}
\usage{
sim.gbm(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values}

\item{model}{contains all the other parameters}

\item{dt}{is the step size}
}
\description{
Simulate from p(X_t|X_{t-1})
Use log-normal transition density specified by the model
}
