% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simFuncs.R
\name{sim.logOU_Discrete}
\alias{sim.logOU_Discrete}
\title{Simulate from 1-D discretized exponential Ornstein Uhlenbeck process
See Bender (SIFIN 2011). Only works in one dimension}
\usage{
sim.logOU_Discrete(x0, model, dt = model$dt)
}
\arguments{
\item{x0}{is the starting values (matrix of size N x model$dim)}

\item{model}{a list containing all the other parameters, including volatility \code{model$sigma},
interest rate \code{model$r} and continuous dividend yield \code{model$div}.}

\item{dt}{is the step size in time. Defaults to \code{model$dt}}
}
\description{
Simulate from 1-D discretized exponential Ornstein Uhlenbeck process
See Bender (SIFIN 2011). Only works in one dimension
}
\details{
Requires
\itemize{
\item  \code{model$rho} -- similar to mean-reversion rate, should be close to 1
\item \code{model$mu} -- mean-reversion level
\item \code{model$sigma} -- volatility
}
}
